Deduplicate IndexedDB RPC errors and split large client transports.
Centralize gRPC-to-sentinel mapping in indexeddb.RPCError, decompose hostindexeddb and gestaltd remote clients under 400 lines per file, drop no-op capability aliases, and simplify S3 ClientError to gRPC codes only.
