+++ sdk/go/internal/hostindexeddb/host_client.go
@@ -3,17 +3,11 @@ package hostindexeddb
 import (
 	"context"
 	"fmt"
-	"strings"
-	"sync"
 
-	. "github.com/valon-technologies/gestalt/sdk/go/indexeddb"
+	idb "github.com/valon-technologies/gestalt/sdk/go/indexeddb"
 	proto "github.com/valon-technologies/gestalt/sdk/go/internal/gen/v1"
-	rpcstatus "google.golang.org/genproto/googleapis/rpc/status"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
-// HostClient is the gRPC host-service implementation of Database.
 type HostClient struct {
 	client proto.IndexedDBClient
 }
@@ -22,17 +16,15 @@ type HostClient struct {
 func (db *HostClient) Close() error { return nil }
 
 var (
-	_ Database     = (*HostClient)(nil)
-	_ ObjectStore  = (*ObjectStoreClient)(nil)
-	_ RangeDeleter = (*ObjectStoreClient)(nil)
-	_ Index        = (*IndexClient)(nil)
-	_ MutableIndex = (*IndexClient)(nil)
-	_ Transaction  = (*hostTransaction)(nil)
-	_ Cursor       = (*hostCursor)(nil)
+	_ idb.Database    = (*HostClient)(nil)
+	_ idb.ObjectStore = (*ObjectStoreClient)(nil)
+	_ idb.Index       = (*IndexClient)(nil)
+	_ idb.Transaction = (*hostTransaction)(nil)
+	_ idb.Cursor      = (*hostCursor)(nil)
 )
 
 // CreateObjectStore creates a named object store with the supplied schema.
-func (db *HostClient) CreateObjectStore(ctx context.Context, name string, schema ObjectStoreOptions) (ObjectStore, error) {
+func (db *HostClient) CreateObjectStore(ctx context.Context, name string, schema idb.ObjectStoreOptions) (idb.ObjectStore, error) {
 	indexes := make([]*proto.IndexSchema, len(schema.Indexes))
 	for i, idx := range schema.Indexes {
 		indexes[i] = &proto.IndexSchema{Name: idx.Name, KeyPath: idx.KeyPath, Unique: idx.Unique}
@@ -63,15 +55,15 @@ func (db *HostClient) DeleteObjectStore(ctx context.Context, name string) error 
 }
 
 // ObjectStore returns a typed handle for working with one object store.
-func (db *HostClient) ObjectStore(name string) ObjectStore {
+func (db *HostClient) ObjectStore(name string) idb.ObjectStore {
 	return &ObjectStoreClient{client: db.client, store: name}
 }
 
 // Transaction starts an explicit IndexedDB transaction over the supplied object
 // store scope.
-func (db *HostClient) Transaction(ctx context.Context, stores []string, mode TransactionMode, opts TransactionOptions) (Transaction, error) {
-	if mode == TransactionVersionChange {
-		return nil, fmt.Errorf("%w: versionchange transactions", ErrUnsupported)
+func (db *HostClient) Transaction(ctx context.Context, stores []string, mode idb.TransactionMode, opts idb.TransactionOptions) (idb.Transaction, error) {
+	if mode == idb.TransactionVersionChange {
+		return nil, fmt.Errorf("%w: versionchange transactions", idb.ErrUnsupported)
 	}
 	if ctx == nil {
 		ctx = context.Background()
@@ -115,7 +107,7 @@ type ObjectStoreClient struct {
 }
 
 // Get loads one record by primary key.
-func (o *ObjectStoreClient) Get(ctx context.Context, id string) (Record, error) {
+func (o *ObjectStoreClient) Get(ctx context.Context, id string) (idb.Record, error) {
 	resp, err := o.client.Get(ctx, &proto.ObjectStoreRequest{Store: o.store, Id: id})
 	if err != nil {
 		return nil, grpcErr(err)
@@ -137,7 +129,7 @@ func (o *ObjectStoreClient) GetKey(ctx context.Context, id string) (string, erro
 }
 
 // Add inserts a new record and fails if its primary key already exists.
-func (o *ObjectStoreClient) Add(ctx context.Context, record Record) error {
+func (o *ObjectStoreClient) Add(ctx context.Context, record idb.Record) error {
 	pbRecord, err := recordToProto(record)
 	if err != nil {
 		return fmt.Errorf("marshal record: %w", err)
@@ -147,7 +139,7 @@ func (o *ObjectStoreClient) Add(ctx context.Context, record Record) error {
 }
 
 // Put upserts a record by primary key.
-func (o *ObjectStoreClient) Put(ctx context.Context, record Record) error {
+func (o *ObjectStoreClient) Put(ctx context.Context, record idb.Record) error {
 	pbRecord, err := recordToProto(record)
 	if err != nil {
 		return fmt.Errorf("marshal record: %w", err)
@@ -169,7 +161,7 @@ func (o *ObjectStoreClient) Clear(ctx context.Context) error {
 }
 
 // GetAll loads all records that match r.
-func (o *ObjectStoreClient) GetAll(ctx context.Context, r *KeyRange) ([]Record, error) {
+func (o *ObjectStoreClient) GetAll(ctx context.Context, r *idb.KeyRange) ([]idb.Record, error) {
 	kr, err := krToProto(r)
 	if err != nil {
 		return nil, err
@@ -186,7 +178,7 @@ func (o *ObjectStoreClient) GetAll(ctx context.Context, r *KeyRange) ([]Record, 
 }
 
 // GetAllKeys loads the primary keys for all records that match r.
-func (o *ObjectStoreClient) GetAllKeys(ctx context.Context, r *KeyRange) ([]string, error) {
+func (o *ObjectStoreClient) GetAllKeys(ctx context.Context, r *idb.KeyRange) ([]string, error) {
 	kr, err := krToProto(r)
 	if err != nil {
 		return nil, err
@@ -199,7 +191,7 @@ func (o *ObjectStoreClient) GetAllKeys(ctx context.Context, r *KeyRange) ([]stri
 }
 
 // Count returns the number of records that match r.
-func (o *ObjectStoreClient) Count(ctx context.Context, r *KeyRange) (int64, error) {
+func (o *ObjectStoreClient) Count(ctx context.Context, r *idb.KeyRange) (int64, error) {
 	kr, err := krToProto(r)
 	if err != nil {
 		return 0, err
@@ -213,7 +205,7 @@ func (o *ObjectStoreClient) Count(ctx context.Context, r *KeyRange) (int64, erro
 
 // DeleteRange removes all records that match r and reports how many were
 // deleted.
-func (o *ObjectStoreClient) DeleteRange(ctx context.Context, r KeyRange) (int64, error) {
+func (o *ObjectStoreClient) DeleteRange(ctx context.Context, r idb.KeyRange) (int64, error) {
 	kr, err := krToProto(&r)
 	if err != nil {
 		return 0, err
@@ -226,17 +218,17 @@ func (o *ObjectStoreClient) DeleteRange(ctx context.Context, r KeyRange) (int64,
 }
 
 // OpenCursor opens a full-value cursor over the object store.
-func (o *ObjectStoreClient) OpenCursor(ctx context.Context, r *KeyRange, dir CursorDirection) (Cursor, error) {
+func (o *ObjectStoreClient) OpenCursor(ctx context.Context, r *idb.KeyRange, dir idb.CursorDirection) (idb.Cursor, error) {
 	return openCursor(ctx, o.client, o.store, "", r, dir, false, nil)
 }
 
 // OpenKeyCursor opens a key-only cursor over the object store.
-func (o *ObjectStoreClient) OpenKeyCursor(ctx context.Context, r *KeyRange, dir CursorDirection) (Cursor, error) {
+func (o *ObjectStoreClient) OpenKeyCursor(ctx context.Context, r *idb.KeyRange, dir idb.CursorDirection) (idb.Cursor, error) {
 	return openCursor(ctx, o.client, o.store, "", r, dir, true, nil)
 }
 
 // Index returns a typed handle for a secondary index on the object store.
-func (o *ObjectStoreClient) Index(name string) Index {
+func (o *ObjectStoreClient) Index(name string) idb.Index {
 	return &IndexClient{client: o.client, store: o.store, index: name}
 }
 
@@ -248,7 +240,7 @@ type IndexClient struct {
 }
 
 // Get loads the first record that matches the supplied index key.
-func (idx *IndexClient) Get(ctx context.Context, values ...any) (Record, error) {
+func (idx *IndexClient) Get(ctx context.Context, values ...any) (idb.Record, error) {
 	vals, err := anyToProtoValues(values)
 	if err != nil {
 		return nil, err
@@ -282,7 +274,7 @@ func (idx *IndexClient) GetKey(ctx context.Context, values ...any) (string, erro
 }
 
 // GetAll loads every record that matches values and r.
-func (idx *IndexClient) GetAll(ctx context.Context, r *KeyRange, values ...any) ([]Record, error) {
+func (idx *IndexClient) GetAll(ctx context.Context, r *idb.KeyRange, values ...any) ([]idb.Record, error) {
 	vals, err := anyToProtoValues(values)
 	if err != nil {
 		return nil, err
@@ -305,7 +297,7 @@ func (idx *IndexClient) GetAll(ctx context.Context, r *KeyRange, values ...any) 
 }
 
 // GetAllKeys loads every primary key that matches values and r.
-func (idx *IndexClient) GetAllKeys(ctx context.Context, r *KeyRange, values ...any) ([]string, error) {
+func (idx *IndexClient) GetAllKeys(ctx context.Context, r *idb.KeyRange, values ...any) ([]string, error) {
 	vals, err := anyToProtoValues(values)
 	if err != nil {
 		return nil, err
@@ -324,7 +316,7 @@ func (idx *IndexClient) GetAllKeys(ctx context.Context, r *KeyRange, values ...a
 }
 
 // Count returns the number of rows that match values and r.
-func (idx *IndexClient) Count(ctx context.Context, r *KeyRange, values ...any) (int64, error) {
+func (idx *IndexClient) Count(ctx context.Context, r *idb.KeyRange, values ...any) (int64, error) {
 	vals, err := anyToProtoValues(values)
 	if err != nil {
 		return 0, err
@@ -348,7 +340,7 @@ func (idx *IndexClient) Delete(ctx context.Context, values ...any) (int64, error
 }
 
 // DeleteRange removes all rows that match values and r.
-func (idx *IndexClient) DeleteRange(ctx context.Context, r *KeyRange, values ...any) (int64, error) {
+func (idx *IndexClient) DeleteRange(ctx context.Context, r *idb.KeyRange, values ...any) (int64, error) {
 	vals, err := anyToProtoValues(values)
 	if err != nil {
 		return 0, err
@@ -367,788 +359,14 @@ func (idx *IndexClient) DeleteRange(ctx context.Context, r *KeyRange, values ...
 }
 
 // OpenCursor opens a full-value cursor over one secondary index.
-func (idx *IndexClient) OpenCursor(ctx context.Context, r *KeyRange, dir CursorDirection, values ...any) (Cursor, error) {
+func (idx *IndexClient) OpenCursor(ctx context.Context, r *idb.KeyRange, dir idb.CursorDirection, values ...any) (idb.Cursor, error) {
 	return openCursor(ctx, idx.client, idx.store, idx.index, r, dir, false, values)
 }
 
 // OpenKeyCursor opens a key-only cursor over one secondary index.
-func (idx *IndexClient) OpenKeyCursor(ctx context.Context, r *KeyRange, dir CursorDirection, values ...any) (Cursor, error) {
+func (idx *IndexClient) OpenKeyCursor(ctx context.Context, r *idb.KeyRange, dir idb.CursorDirection, values ...any) (idb.Cursor, error) {
 	return openCursor(ctx, idx.client, idx.store, idx.index, r, dir, true, values)
 }
 
 // HostTransaction is the host-service transaction implementation.
 type HostTransaction = hostTransaction
-
-// hostTransaction is an explicit IndexedDB transaction over a fixed store scope.
-type hostTransaction struct {
-	stream proto.IndexedDB_TransactionClient
-	cancel context.CancelFunc
-	mu     sync.Mutex
-	nextID uint64
-	done   bool
-	err    error
-}
-
-// ObjectStore returns a transaction-scoped object store handle.
-func (tx *hostTransaction) ObjectStore(name string) TransactionObjectStore {
-	return &hostTxObjectStore{tx: tx, store: name}
-}
-
-// Commit atomically commits all writes made in the transaction.
-func (tx *hostTransaction) Commit(ctx context.Context) error {
-	_ = ctx
-	tx.mu.Lock()
-	if tx.done {
-		err := tx.err
-		tx.mu.Unlock()
-		if err != nil {
-			return err
-		}
-		return ErrTransactionDone
-	}
-	if tx.err != nil {
-		err := tx.err
-		tx.mu.Unlock()
-		return err
-	}
-	tx.done = true
-
-	if err := tx.stream.Send(&proto.TransactionClientMessage{Msg: &proto.TransactionClientMessage_Commit{Commit: &proto.TransactionCommitRequest{}}}); err != nil {
-		return tx.failLocked(grpcErr(err))
-	}
-	resp, err := tx.stream.Recv()
-	if err != nil {
-		return tx.failLocked(grpcErr(err))
-	}
-	commit := resp.GetCommit()
-	if commit == nil {
-		return tx.failLocked(fmt.Errorf("indexeddb: expected transaction commit response"))
-	}
-	if err := rpcStatusErr(commit.GetError()); err != nil {
-		return tx.failLocked(err)
-	}
-	tx.mu.Unlock()
-	tx.cleanup()
-	return nil
-}
-
-// Abort rolls back the transaction.
-func (tx *hostTransaction) Abort(ctx context.Context) error {
-	_ = ctx
-	tx.mu.Lock()
-	if tx.done {
-		err := tx.err
-		tx.mu.Unlock()
-		if err != nil {
-			return err
-		}
-		return ErrTransactionDone
-	}
-	tx.done = true
-
-	if err := tx.stream.Send(&proto.TransactionClientMessage{Msg: &proto.TransactionClientMessage_Abort{Abort: &proto.TransactionAbortRequest{}}}); err != nil {
-		return tx.failLocked(grpcErr(err))
-	}
-	resp, err := tx.stream.Recv()
-	if err != nil {
-		return tx.failLocked(grpcErr(err))
-	}
-	abort := resp.GetAbort()
-	if abort == nil {
-		return tx.failLocked(fmt.Errorf("indexeddb: expected transaction abort response"))
-	}
-	if err := rpcStatusErr(abort.GetError()); err != nil {
-		return tx.failLocked(err)
-	}
-	tx.mu.Unlock()
-	tx.cleanup()
-	return nil
-}
-
-func (tx *hostTransaction) sendOperation(op *proto.TransactionOperation) (*proto.TransactionOperationResponse, error) {
-	tx.mu.Lock()
-	if tx.done {
-		err := tx.err
-		tx.mu.Unlock()
-		if err != nil {
-			return nil, err
-		}
-		return nil, ErrTransactionDone
-	}
-	if tx.err != nil {
-		err := tx.err
-		tx.mu.Unlock()
-		return nil, err
-	}
-	tx.nextID++
-	op.RequestId = tx.nextID
-
-	if err := tx.stream.Send(&proto.TransactionClientMessage{Msg: &proto.TransactionClientMessage_Operation{Operation: op}}); err != nil {
-		return nil, tx.failLocked(grpcErr(err))
-	}
-	resp, err := tx.stream.Recv()
-	if err != nil {
-		return nil, tx.failLocked(grpcErr(err))
-	}
-	opResp := resp.GetOperation()
-	if opResp == nil {
-		return nil, tx.failLocked(fmt.Errorf("indexeddb: expected transaction operation response"))
-	}
-	if opResp.GetRequestId() != op.GetRequestId() {
-		return nil, tx.failLocked(fmt.Errorf("indexeddb: response request id %d does not match %d", opResp.GetRequestId(), op.GetRequestId()))
-	}
-	if err := rpcStatusErr(opResp.GetError()); err != nil {
-		tx.done = true
-		tx.err = err
-		tx.mu.Unlock()
-		tx.cleanup()
-		return nil, err
-	}
-	tx.mu.Unlock()
-	return opResp, nil
-}
-
-func (tx *hostTransaction) failLocked(err error) error {
-	tx.err = err
-	tx.done = true
-	tx.mu.Unlock()
-	tx.cleanup()
-	return err
-}
-
-func (tx *hostTransaction) cleanup() {
-	if tx.stream != nil {
-		_ = tx.stream.CloseSend()
-		tx.stream = nil
-	}
-	if tx.cancel != nil {
-		tx.cancel()
-		tx.cancel = nil
-	}
-}
-
-// TransactionObjectStore provides transaction-scoped object-store operations.
-type hostTxObjectStore struct {
-	tx    *hostTransaction
-	store string
-}
-
-func (s *hostTxObjectStore) Get(ctx context.Context, id string) (Record, error) {
-	_ = ctx
-	resp, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_Get{Get: &proto.ObjectStoreRequest{Store: s.store, Id: id}}})
-	if err != nil {
-		return nil, err
-	}
-	record, err := recordFromProto(resp.GetRecord().GetRecord())
-	if err != nil {
-		return nil, fmt.Errorf("unmarshal record: %w", err)
-	}
-	return record, nil
-}
-
-func (s *hostTxObjectStore) GetKey(ctx context.Context, id string) (string, error) {
-	_ = ctx
-	resp, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_GetKey{GetKey: &proto.ObjectStoreRequest{Store: s.store, Id: id}}})
-	if err != nil {
-		return "", err
-	}
-	return resp.GetKey().GetKey(), nil
-}
-
-func (s *hostTxObjectStore) Add(ctx context.Context, record Record) error {
-	_ = ctx
-	pbRecord, err := recordToProto(record)
-	if err != nil {
-		return fmt.Errorf("marshal record: %w", err)
-	}
-	_, err = s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_Add{Add: &proto.RecordRequest{Store: s.store, Record: pbRecord}}})
-	return err
-}
-
-func (s *hostTxObjectStore) Put(ctx context.Context, record Record) error {
-	_ = ctx
-	pbRecord, err := recordToProto(record)
-	if err != nil {
-		return fmt.Errorf("marshal record: %w", err)
-	}
-	_, err = s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_Put{Put: &proto.RecordRequest{Store: s.store, Record: pbRecord}}})
-	return err
-}
-
-func (s *hostTxObjectStore) Delete(ctx context.Context, id string) error {
-	_ = ctx
-	_, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_Delete{Delete: &proto.ObjectStoreRequest{Store: s.store, Id: id}}})
-	return err
-}
-
-func (s *hostTxObjectStore) Clear(ctx context.Context) error {
-	_ = ctx
-	_, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_Clear{Clear: &proto.ObjectStoreNameRequest{Store: s.store}}})
-	return err
-}
-
-func (s *hostTxObjectStore) GetAll(ctx context.Context, r *KeyRange) ([]Record, error) {
-	_ = ctx
-	kr, err := krToProto(r)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_GetAll{GetAll: &proto.ObjectStoreRangeRequest{Store: s.store, Range: kr}}})
-	if err != nil {
-		return nil, err
-	}
-	records, err := recordsFromProto(resp.GetRecords().GetRecords())
-	if err != nil {
-		return nil, fmt.Errorf("unmarshal records: %w", err)
-	}
-	return records, nil
-}
-
-func (s *hostTxObjectStore) GetAllKeys(ctx context.Context, r *KeyRange) ([]string, error) {
-	_ = ctx
-	kr, err := krToProto(r)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_GetAllKeys{GetAllKeys: &proto.ObjectStoreRangeRequest{Store: s.store, Range: kr}}})
-	if err != nil {
-		return nil, err
-	}
-	return resp.GetKeys().GetKeys(), nil
-}
-
-func (s *hostTxObjectStore) Count(ctx context.Context, r *KeyRange) (int64, error) {
-	_ = ctx
-	kr, err := krToProto(r)
-	if err != nil {
-		return 0, err
-	}
-	resp, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_Count{Count: &proto.ObjectStoreRangeRequest{Store: s.store, Range: kr}}})
-	if err != nil {
-		return 0, err
-	}
-	return resp.GetCount().GetCount(), nil
-}
-
-func (s *hostTxObjectStore) DeleteRange(ctx context.Context, r KeyRange) (int64, error) {
-	_ = ctx
-	kr, err := krToProto(&r)
-	if err != nil {
-		return 0, err
-	}
-	resp, err := s.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_DeleteRange{DeleteRange: &proto.ObjectStoreRangeRequest{Store: s.store, Range: kr}}})
-	if err != nil {
-		return 0, err
-	}
-	return resp.GetDelete().GetDeleted(), nil
-}
-
-func (s *hostTxObjectStore) Index(name string) TransactionIndex {
-	return &hostTxIndex{tx: s.tx, store: s.store, index: name}
-}
-
-// TransactionIndex provides transaction-scoped index operations.
-type hostTxIndex struct {
-	tx    *hostTransaction
-	store string
-	index string
-}
-
-func (idx *hostTxIndex) Get(ctx context.Context, values ...any) (Record, error) {
-	_ = ctx
-	req, err := idx.query(nil, values)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := idx.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_IndexGet{IndexGet: req}})
-	if err != nil {
-		return nil, err
-	}
-	record, err := recordFromProto(resp.GetRecord().GetRecord())
-	if err != nil {
-		return nil, fmt.Errorf("unmarshal record: %w", err)
-	}
-	return record, nil
-}
-
-func (idx *hostTxIndex) GetKey(ctx context.Context, values ...any) (string, error) {
-	_ = ctx
-	req, err := idx.query(nil, values)
-	if err != nil {
-		return "", err
-	}
-	resp, err := idx.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_IndexGetKey{IndexGetKey: req}})
-	if err != nil {
-		return "", err
-	}
-	return resp.GetKey().GetKey(), nil
-}
-
-func (idx *hostTxIndex) GetAll(ctx context.Context, r *KeyRange, values ...any) ([]Record, error) {
-	_ = ctx
-	req, err := idx.query(r, values)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := idx.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_IndexGetAll{IndexGetAll: req}})
-	if err != nil {
-		return nil, err
-	}
-	records, err := recordsFromProto(resp.GetRecords().GetRecords())
-	if err != nil {
-		return nil, fmt.Errorf("unmarshal records: %w", err)
-	}
-	return records, nil
-}
-
-func (idx *hostTxIndex) GetAllKeys(ctx context.Context, r *KeyRange, values ...any) ([]string, error) {
-	_ = ctx
-	req, err := idx.query(r, values)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := idx.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_IndexGetAllKeys{IndexGetAllKeys: req}})
-	if err != nil {
-		return nil, err
-	}
-	return resp.GetKeys().GetKeys(), nil
-}
-
-func (idx *hostTxIndex) Count(ctx context.Context, r *KeyRange, values ...any) (int64, error) {
-	_ = ctx
-	req, err := idx.query(r, values)
-	if err != nil {
-		return 0, err
-	}
-	resp, err := idx.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_IndexCount{IndexCount: req}})
-	if err != nil {
-		return 0, err
-	}
-	return resp.GetCount().GetCount(), nil
-}
-
-func (idx *hostTxIndex) Delete(ctx context.Context, values ...any) (int64, error) {
-	return idx.DeleteRange(ctx, nil, values...)
-}
-
-// DeleteRange removes all transaction-scoped rows that match values and r.
-func (idx *hostTxIndex) DeleteRange(ctx context.Context, r *KeyRange, values ...any) (int64, error) {
-	_ = ctx
-	req, err := idx.query(r, values)
-	if err != nil {
-		return 0, err
-	}
-	resp, err := idx.tx.sendOperation(&proto.TransactionOperation{Operation: &proto.TransactionOperation_IndexDelete{IndexDelete: req}})
-	if err != nil {
-		return 0, err
-	}
-	return resp.GetDelete().GetDeleted(), nil
-}
-
-func (idx *hostTxIndex) query(r *KeyRange, values []any) (*proto.IndexQueryRequest, error) {
-	vals, err := anyToProtoValues(values)
-	if err != nil {
-		return nil, err
-	}
-	kr, err := krToProto(r)
-	if err != nil {
-		return nil, err
-	}
-	return &proto.IndexQueryRequest{Store: idx.store, Index: idx.index, Values: vals, Range: kr}, nil
-}
-
-// Cursor streams IndexedDB rows one at a time.
-// HostCursor is the host-service cursor implementation.
-type HostCursor = hostCursor
-
-type hostCursor struct {
-	stream      proto.IndexedDB_OpenCursorClient
-	cancel      context.CancelFunc
-	keysOnly    bool
-	indexCursor bool
-	entry       *proto.CursorEntry
-	err         error
-	done        bool
-}
-
-// Continue advances the cursor by one row.
-func (c *hostCursor) Continue() bool {
-	return c.sendAndRecv(&proto.CursorCommand{
-		Command: &proto.CursorCommand_Next{Next: true},
-	})
-}
-
-// ContinueToKey advances the cursor to the supplied key, or exhausts it if the
-// key does not exist.
-func (c *hostCursor) ContinueToKey(key any) bool {
-	kvs, err := cursorKeyToProto(key, c.indexCursor)
-	if err != nil {
-		c.err = err
-		return false
-	}
-	return c.sendAndRecv(&proto.CursorCommand{
-		Command: &proto.CursorCommand_ContinueToKey{ContinueToKey: &proto.CursorKeyTarget{Key: kvs}},
-	})
-}
-
-// Advance skips count rows ahead.
-func (c *hostCursor) Advance(count int) bool {
-	return c.sendAndRecv(&proto.CursorCommand{
-		Command: &proto.CursorCommand_Advance{Advance: int32(count)},
-	})
-}
-
-// Key returns the current cursor key.
-func (c *hostCursor) Key() any {
-	if c.entry == nil || len(c.entry.GetKey()) == 0 {
-		return nil
-	}
-	parts, err := keyValuesToAny(c.entry.GetKey())
-	if err != nil {
-		c.err = err
-		return nil
-	}
-	if !c.indexCursor && len(parts) == 1 {
-		return parts[0]
-	}
-	return parts
-}
-
-// PrimaryKey returns the current record's primary key.
-func (c *hostCursor) PrimaryKey() string {
-	if c.entry == nil {
-		return ""
-	}
-	return c.entry.GetPrimaryKey()
-}
-
-// Value returns the current record.
-func (c *hostCursor) Value() (Record, error) {
-	if c.keysOnly {
-		return nil, ErrKeysOnly
-	}
-	if c.entry == nil || c.entry.GetRecord() == nil {
-		return nil, ErrNotFound
-	}
-	return recordFromProto(c.entry.GetRecord())
-}
-
-// Delete removes the current row and keeps the cursor open.
-func (c *hostCursor) Delete() error {
-	if c.err != nil {
-		return c.err
-	}
-	if c.done {
-		return ErrNotFound
-	}
-	err := c.stream.Send(&proto.CursorClientMessage{
-		Msg: &proto.CursorClientMessage_Command{
-			Command: &proto.CursorCommand{
-				Command: &proto.CursorCommand_Delete{Delete: true},
-			},
-		},
-	})
-	if err != nil {
-		return c.setErr(grpcErr(err))
-	}
-	resp, err := c.stream.Recv()
-	if err != nil {
-		return c.setErr(grpcErr(err))
-	}
-	if resp == nil {
-		return c.setErr(fmt.Errorf("indexeddb: cursor stream ended during mutation"))
-	}
-	switch v := resp.GetResult().(type) {
-	case *proto.CursorResponse_Entry:
-		c.entry = v.Entry
-	case *proto.CursorResponse_Done:
-		if v.Done {
-			c.done = true
-			c.entry = nil
-		}
-	default:
-		return c.setErr(fmt.Errorf("indexeddb: unexpected cursor mutation ack"))
-	}
-	return nil
-}
-
-// Update replaces the current row and keeps the cursor open.
-func (c *hostCursor) Update(value Record) error {
-	if c.err != nil {
-		return c.err
-	}
-	if c.done {
-		return ErrNotFound
-	}
-	pbRecord, err := recordToProto(value)
-	if err != nil {
-		return fmt.Errorf("indexeddb: marshal cursor update: %w", err)
-	}
-	err = c.stream.Send(&proto.CursorClientMessage{
-		Msg: &proto.CursorClientMessage_Command{
-			Command: &proto.CursorCommand{
-				Command: &proto.CursorCommand_Update{Update: pbRecord},
-			},
-		},
-	})
-	if err != nil {
-		return c.setErr(grpcErr(err))
-	}
-	resp, err := c.stream.Recv()
-	if err != nil {
-		return c.setErr(grpcErr(err))
-	}
-	if resp == nil {
-		return c.setErr(fmt.Errorf("indexeddb: cursor stream ended during mutation"))
-	}
-	switch v := resp.GetResult().(type) {
-	case *proto.CursorResponse_Entry:
-		c.entry = v.Entry
-	case *proto.CursorResponse_Done:
-		if v.Done {
-			c.done = true
-			c.entry = nil
-		} else if c.entry != nil {
-			c.entry.Record = pbRecord
-		}
-	default:
-		return c.setErr(fmt.Errorf("indexeddb: unexpected cursor mutation ack"))
-	}
-	return nil
-}
-
-// Err returns the terminal cursor error, if any.
-func (c *hostCursor) Err() error {
-	return c.err
-}
-
-func (c *hostCursor) cleanup() error {
-	var err error
-	if c.stream != nil {
-		err = grpcErr(c.stream.CloseSend())
-		c.stream = nil
-	}
-	if c.cancel != nil {
-		c.cancel()
-		c.cancel = nil
-	}
-	return err
-}
-
-func (c *hostCursor) setErr(err error) error {
-	c.err = err
-	_ = c.cleanup()
-	return c.err
-}
-
-// Close closes the cursor stream and releases its transport resources.
-func (c *hostCursor) Close() error {
-	c.done = true
-	c.entry = nil
-	if c.stream == nil {
-		return c.cleanup()
-	}
-	sendErr := c.stream.Send(&proto.CursorClientMessage{
-		Msg: &proto.CursorClientMessage_Command{
-			Command: &proto.CursorCommand{
-				Command: &proto.CursorCommand_Close{Close: true},
-			},
-		},
-	})
-	closeErr := c.cleanup()
-	if sendErr != nil {
-		return grpcErr(sendErr)
-	}
-	return closeErr
-}
-
-func (c *hostCursor) sendAndRecv(cmd *proto.CursorCommand) bool {
-	if c.done || c.err != nil {
-		return false
-	}
-	err := c.stream.Send(&proto.CursorClientMessage{
-		Msg: &proto.CursorClientMessage_Command{Command: cmd},
-	})
-	if err != nil {
-		_ = c.setErr(grpcErr(err))
-		return false
-	}
-	resp, err := c.stream.Recv()
-	if err != nil {
-		_ = c.setErr(grpcErr(err))
-		return false
-	}
-	if resp == nil {
-		_ = c.setErr(fmt.Errorf("indexeddb: cursor stream ended"))
-		return false
-	}
-	switch v := resp.GetResult().(type) {
-	case *proto.CursorResponse_Entry:
-		c.entry = v.Entry
-		return true
-	case *proto.CursorResponse_Done:
-		if !v.Done {
-			_ = c.setErr(fmt.Errorf("indexeddb: unexpected non-exhaustion cursor ack"))
-			c.entry = nil
-			return false
-		}
-		c.done = true
-		c.entry = nil
-		return false
-	default:
-		_ = c.setErr(fmt.Errorf("indexeddb: unexpected cursor response"))
-		c.entry = nil
-		return false
-	}
-}
-
-func cursorDirectionToProto(dir CursorDirection) proto.CursorDirection {
-	switch dir {
-	case CursorNextUnique:
-		return proto.CursorDirection_CURSOR_NEXT_UNIQUE
-	case CursorPrev:
-		return proto.CursorDirection_CURSOR_PREV
-	case CursorPrevUnique:
-		return proto.CursorDirection_CURSOR_PREV_UNIQUE
-	default:
-		return proto.CursorDirection_CURSOR_NEXT
-	}
-}
-
-func openCursor(ctx context.Context, client proto.IndexedDBClient, store, index string, r *KeyRange, dir CursorDirection, keysOnly bool, values []any) (Cursor, error) {
-	kr, err := krToProto(r)
-	if err != nil {
-		return nil, err
-	}
-	vals, err := typedValuesFromAny(values)
-	if err != nil {
-		return nil, err
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	streamCtx, streamCancel := context.WithCancel(ctx)
-	stream, err := client.OpenCursor(streamCtx)
-	if err != nil {
-		streamCancel()
-		return nil, grpcErr(err)
-	}
-	err = stream.Send(&proto.CursorClientMessage{
-		Msg: &proto.CursorClientMessage_Open{
-			Open: &proto.OpenCursorRequest{
-				Store:     store,
-				Range:     kr,
-				Direction: cursorDirectionToProto(dir),
-				KeysOnly:  keysOnly,
-				Index:     index,
-				Values:    vals,
-			},
-		},
-	})
-	if err != nil {
-		_ = stream.CloseSend()
-		streamCancel()
-		return nil, grpcErr(err)
-	}
-	// Read the open ack to surface creation errors synchronously.
-	resp, err := stream.Recv()
-	if err != nil {
-		_ = stream.CloseSend()
-		streamCancel()
-		return nil, grpcErr(err)
-	}
-	if resp == nil {
-		_ = stream.CloseSend()
-		streamCancel()
-		return nil, fmt.Errorf("indexeddb: cursor stream ended during open")
-	}
-	done, ok := resp.GetResult().(*proto.CursorResponse_Done)
-	if !ok || done.Done {
-		_ = stream.CloseSend()
-		streamCancel()
-		return nil, fmt.Errorf("indexeddb: unexpected cursor open ack")
-	}
-	return &hostCursor{stream: stream, cancel: streamCancel, keysOnly: keysOnly, indexCursor: index != ""}, nil
-}
-
-func krToProto(r *KeyRange) (*proto.KeyRange, error) {
-	if r == nil {
-		return nil, nil
-	}
-	kr := &proto.KeyRange{LowerOpen: r.LowerOpen, UpperOpen: r.UpperOpen}
-	if r.Lower != nil {
-		v, err := typedValueFromAny(r.Lower)
-		if err != nil {
-			return nil, fmt.Errorf("marshal key range lower: %w", err)
-		}
-		kr.Lower = v
-	}
-	if r.Upper != nil {
-		v, err := typedValueFromAny(r.Upper)
-		if err != nil {
-			return nil, fmt.Errorf("marshal key range upper: %w", err)
-		}
-		kr.Upper = v
-	}
-	return kr, nil
-}
-
-func anyToProtoValues(values []any) ([]*proto.TypedValue, error) {
-	return typedValuesFromAny(values)
-}
-
-func transactionModeToProto(mode TransactionMode) proto.TransactionMode {
-	if mode == TransactionReadwrite {
-		return proto.TransactionMode_TRANSACTION_READWRITE
-	}
-	return proto.TransactionMode_TRANSACTION_READONLY
-}
-
-func durabilityHintToProto(hint TransactionDurabilityHint) proto.TransactionDurabilityHint {
-	switch hint {
-	case TransactionDurabilityStrict:
-		return proto.TransactionDurabilityHint_TRANSACTION_DURABILITY_STRICT
-	case TransactionDurabilityRelaxed:
-		return proto.TransactionDurabilityHint_TRANSACTION_DURABILITY_RELAXED
-	default:
-		return proto.TransactionDurabilityHint_TRANSACTION_DURABILITY_DEFAULT
-	}
-}
-
-func rpcStatusErr(st *rpcstatus.Status) error {
-	if st == nil || st.GetCode() == int32(codes.OK) {
-		return nil
-	}
-	return grpcErr(status.Error(codes.Code(st.GetCode()), st.GetMessage()))
-}
-
-func grpcErr(err error) error {
-	if err == nil {
-		return nil
-	}
-	st, ok := status.FromError(err)
-	if !ok {
-		return err
-	}
-	switch st.Code() {
-	case codes.NotFound:
-		return ErrNotFound
-	case codes.AlreadyExists:
-		return ErrAlreadyExists
-	case codes.InvalidArgument:
-		if strings.Contains(st.Message(), "invalid transaction") {
-			return ErrInvalidTransaction
-		}
-		return err
-	case codes.FailedPrecondition:
-		if strings.Contains(st.Message(), "readonly") {
-			return ErrReadOnly
-		}
-		if strings.Contains(st.Message(), "already finished") {
-			return ErrTransactionDone
-		}
-		return err
-	default:
-		return err
-	}
-}
