+++ sdk/go/s3/client_errors.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-// ClientError maps provider, transport, and S3-compatible service errors to S3 client sentinel errors.
+// ClientError maps provider and transport gRPC errors to S3 client sentinel errors.
 func ClientError(err error) error {
 	if err == nil {
 		return nil
@@ -19,20 +19,8 @@ func ClientError(err error) error {
 	if mapped := rpcClientError(err); mapped != err {
 		return mapped
 	}
-	if isS3NotFound(err) {
-		return ErrNotFound
-	}
-	if isS3PreconditionFailed(err) {
-		return ErrPreconditionFailed
-	}
-	if isS3InvalidRange(err) {
-		return ErrInvalidRange
-	}
 	if strings.Contains(err.Error(), "expires must be") {
 		return status.Error(codes.InvalidArgument, err.Error())
-	}
-	if strings.Contains(err.Error(), "s3: invalid range") {
-		return ErrInvalidRange
 	}
 	return err
 }
@@ -52,27 +40,4 @@ func rpcClientError(err error) error {
 	default:
 		return err
 	}
-}
-
-func isS3NotFound(err error) bool {
-	msg := err.Error()
-	return strings.Contains(msg, "NoSuchKey") ||
-		strings.Contains(msg, "NoSuchBucket") ||
-		strings.Contains(msg, "NoSuchVersion") ||
-		strings.Contains(msg, "StatusCode: 404")
-}
-
-func isS3PreconditionFailed(err error) bool {
-	msg := err.Error()
-	return strings.Contains(msg, "PreconditionFailed") ||
-		strings.Contains(msg, "NotModified") ||
-		strings.Contains(msg, "StatusCode: 412") ||
-		strings.Contains(msg, "StatusCode: 304")
-}
-
-func isS3InvalidRange(err error) bool {
-	msg := err.Error()
-	return strings.Contains(msg, "s3: invalid range") ||
-		strings.Contains(msg, "InvalidRange") ||
-		strings.Contains(msg, "StatusCode: 416")
 }
